pyproject.toml
[project]
name = "md2docx"
version = "0.1.0"
description = "Map Markdown sections into a Word template by heading."
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
  "fastapi>=0.116.0,<1.0.0",
  "markdown-it-py>=3.0.0,<4.0.0",
  "python-docx>=1.1.0,<2.0.0",
  "python-multipart>=0.0.9,<1.0.0",
  "PyYAML>=6.0.0,<7.0.0",
  "typer>=0.12.0,<1.0.0",
  "uvicorn[standard]>=0.35.0,<1.0.0",
]

[project.scripts]
md2docx = "md2docx.cli:app"

[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[tool.setuptools]
package-dir = {"" = "src"}

[tool.setuptools.packages.find]
where = ["src"]

[tool.pytest.ini_options]
pythonpath = ["src"]
